Split SearchResult from generic classes
